Add files via upload
This project was developed under MotionCut java programming internship and contains a simple Calculator Program which is capable of taking input from the user and do the desired operation and show the result to the user as well. (The jre and setup installation files are not included in this project because of the size limit.). The Program can directly using the task1.jar or Calculator.java in a java compiler with jre above than 1.4.